Fetch current user through API
This commit updates how the current user and their access token are
passed into the front end.

Instead of passing in an access token from a hard-coded environment
variable, the access token is now retrieved from the current user. This
is facilitated by a new sign-in-protected route that returns the current
user when present.

On the front end, two new variants were added to the Model: Loading and
SignIn.

Loading is used as a transition state when various API requests are
being made.

SignIn is the initial state of the application. The model is also set to
SignIn when actions that required a current user take place. When the
Model is set to SignIn, the sign in link to GroupMe is displayed.