Add Message model and code for fetching messages
This PR introduces some changes which lay the foundation for caching
messages in our database. Future work will set up a way to trigger and
managed the message fetching process. We'll likely offload this process
to background job.

There's a new `GroupMe::FetchMessages` class which follows a similar
pattern to the previously added `GroupMe::FetchGroups`. One thing not
yet handled in this new class is how to deal with rate-limiting in the
GroupMe API. We'll probably need to add some pause between pages.

In the migration, we've created a `raw_message` JSON column to store the
raw message data. If we ever want to add new columns to the db, we'll be
able to fetch that information from that JSON column.