Introduce MessageCache model and MessageCacheJob
This PR adds the ability for a user to initialize the caching of
messages from the GroupMe API.

A group can only have one cache job running and can only start one once
every 24 hours. The `MessageCache` model and related table were added
to keep track of when a group's messages are being cached. This new model
renders the `messages_last_cached_at` column on `groups` redundant, so
it was removed.

The GroupMe::FetchGroups class was updated to be able to start fetching
messages from a given message id (`after_id`) if provided. This allows
it to be used for two occasions:
1. No `after_id`: Caching all messages since the start of the group
1. `after_id` provided: Updating cache with new messages since the
   message with the given id was sent

An API endpoint was exposed to kick-off the caching process. It requires
a logged in user who is a member of the group matching the given id.