Hook up most liked messages page
This commit gets the basics for the most liked messages page hooked up.

A user can visit the page from the groups list. By default, the page
tries to load the top 10 most liked messages (messages ordered by
`favorites_count` descending).

The page also tries to load the latest message cache record for the
group. Different messages are shown to the user depending on whether or
not a message cache record  is present or if the back end is currently
caching messages from the GroupMe API. A button is also shown to allow
the user to re-cache the groups messages.

We introduced a new package that allows us to pull detailed information
about errored HTTP requests. This is useful for us in the case that an
endpoint returns error message text in the body of a non-200 response.
In this commit, that happens in the `create` action in the
`MessageCachesController`.

Follow-up work will sort out the design and styling of the pages.