Create separate services for fetching all vs recent messages
Due to the way GroupMe structures its messages API response, one can
only traverse the message history based on a given id.

For the case that we want to fetch the entire history of messages, we
first need to fetch the most recent messages, and then page back through
the history using the last message of that batch of messages. This is
passed as the `before_id` param in the request url.

For the case where we want to fetch the most recent messages since the
last time we cached them, we want to use the ID of the last message we
fetched as the `after_id`.