Add 'Show More' button to MostLiked page
This commit adds a new button to the bottom of the most liked page that
allows the user to page back through the message history.

When ordering by only `favorites_count`, the results were
non-deterministic. An additional ordering clause (by `created_at`
descending) was added to keep the ordering of the message list
consistent between pages and refreshes.